requirements-api.txt
fastapi
uvicorn[standard]
pydantic
httpx
pandas
numpy
yfinance
openpyxl
pyyaml
python-dotenv
statsmodels
scikit-learn
fredapi
plotly
requests
rich
python-dateutil
edgartools
